+++ src/tools.cpp
@@ -1,8 +1,8 @@
 #include "tools.h"
 #include <iostream>
 
-Eigen::Matrix4f Homogeneous_matrix(Eigen::MatrixXf mat) {
-    Eigen::Matrix4f homo_matrix = Eigen::Matrix4f::Zero();
+Eigen::Matrix4d Homogeneous_matrix(Eigen::MatrixXd mat) {
+    Eigen::Matrix4d homo_matrix = Eigen::Matrix4d::Zero();
     homo_matrix.block<3, 4>(0, 0) = mat;
     homo_matrix(3, 3) = 1;
     return homo_matrix;
